feat: Add jwt-decode library for token decoding
Add jwt-decode library to the project dependencies for decoding JWT tokens. This library will be used to extract the role from the token and store it in the localStorage.

Also, update the axiosInstance to include the token in the request headers. The token is retrieved from the localStorage and added as a Bearer token.

Lastly, update the Create User page to handle errors when creating a user. If the user does not have the necessary permissions, display an error message. Otherwise, display a generic error message.

Closes #123